# openapi.yaml
# This file was autogenerated, please do not modify.
components:
  schemas:
    DocumentType:
      description: Enum for the type of document.
      enum:
      - Dataset
      - Project
      - Study
      - Experiment
      - Sample
      - Publication
      - File
      title: DocumentType
      type: string
    Facet:
      description: Represents a facet and the possible values for that facet.
      properties:
        key:
          title: Key
          type: string
        options:
          items:
            $ref: '#/components/schemas/FacetOption'
          title: Options
          type: array
      required:
      - key
      - options
      title: Facet
      type: object
    FacetOption:
      description: Represent values and their corresponding count for a facet.
      properties:
        count:
          title: Count
          type: integer
        option:
          title: Option
          type: string
      required:
      - option
      title: FacetOption
      type: object
    FilterOption:
      description: Represents a Filter option.
      properties:
        key:
          title: Key
          type: string
        value:
          title: Value
          type: string
      required:
      - key
      - value
      title: FilterOption
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    SearchHit:
      description: Represents the Search Hit.
      properties:
        content:
          title: Content
          type: object
        context:
          title: Context
          type: string
        document_type:
          $ref: '#/components/schemas/DocumentType'
        id:
          title: Id
          type: string
      required:
      - document_type
      - id
      title: SearchHit
      type: object
    SearchQuery:
      description: Represents the Search Query.
      properties:
        filters:
          items:
            $ref: '#/components/schemas/FilterOption'
          title: Filters
          type: array
        query:
          title: Query
          type: string
      required:
      - query
      title: SearchQuery
      type: object
    SearchResult:
      description: Represents the Search Result.
      properties:
        facets:
          items:
            $ref: '#/components/schemas/Facet'
          title: Facets
          type: array
        hits:
          items:
            $ref: '#/components/schemas/SearchHit'
          title: Hits
          type: array
      required:
      - facets
      - hits
      title: SearchResult
      type: object
    ValidationError:
      properties:
        loc:
          items:
            type: string
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
  /:
    get:
      description: Index for Metadata Search Service.
      operationId: index__get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Index for Metadata Search Service
  /rpc/search:
    post:
      description: Search metadata based on a given query string.
      operationId: search_rpc_search_post
      parameters:
      - in: query
        name: document_type
        required: true
        schema:
          $ref: '#/components/schemas/DocumentType'
      - in: query
        name: facet
        required: false
        schema:
          default: false
          title: Facet
          type: boolean
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          title: Limit
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchQuery'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search metadata by keywords and facets